+++ database/migrations/2019_08_17_100000_create_users_table.php
@@ -1,8 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 class CreateUsersTable extends Migration
 {
@@ -13,22 +14,26 @@ class CreateUsersTable extends Migration
    */
   public function up()
   {
-    Schema::create('users', function (Blueprint $table) {
-      $table->id();
-      $table->string('firstName');
-      $table->string('lastName');
-      $table->foreignId('documentType')->constrained('document_types');
-      $table->string('documentNumber')->unique();
-      $table->unsignedSmallInteger('verificationDigit');
-      $table->string('email')->unique();
-      $table->timestamp('email_verified_at')->nullable();
-      $table->string('password');
-      $table->string('mobileNumber');
-      $table->foreignId('category')->constrained('user_categories');
-      $table->rememberToken();
-      $table->timestamps();
-      $table->softDeletes();
-    });
+    if (!Schema::hasTable('users') || Config::get('app.dropUsers', true)) {
+      Schema::create('users', function (Blueprint $table) {
+        $table->id();
+        $table->string('firstName');
+        $table->string('lastName');
+        $table->foreignId('documentType')->constrained('document_types');
+        $table->string('documentNumber')->unique();
+        $table->unsignedSmallInteger('verificationDigit');
+        $table->string('email')->unique();
+        $table->timestamp('email_verified_at')->nullable();
+        $table->tinyInteger('sendEmails')->default(1)->require();
+        $table->string('password');
+        $table->string('mobileNumber');
+        $table->foreignId('category')->constrained('user_categories')->default(1);
+        $table->tinyInteger('active')->default(1)->require();
+        $table->rememberToken();
+        $table->timestamps();
+        $table->softDeletes();
+      });
+    }
   }
 
   /**
@@ -38,6 +43,8 @@ class CreateUsersTable extends Migration
    */
   public function down()
   {
-    Schema::dropIfExists('users');
+    if (Config::get('app.dropUsers', false)) {
+      Schema::dropIfExists('users');
+    }
   }
 }
